+++ neoforge/src/main/java/de/cech12/unlitcampfire/platform/NeoForgeConfigHelper.java
@@ -25,7 +25,6 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     public static final ModConfigSpec.IntValue CAMPFIRE_RUN_OUT_INDICATOR_TIME;
     public static final ModConfigSpec.IntValue CAMPFIRE_RAIN_UNLIT_TIME;
     public static final ModConfigSpec.IntValue CAMPFIRE_RAIN_PARTICLE_FACTOR;
-    public static final ModConfigSpec.BooleanValue CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN;
     public static final ModConfigSpec.BooleanValue CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME;
     public static final ModConfigSpec.BooleanValue CAMPFIRE_ADDING_BURNABLES;
     public static final ModConfigSpec.IntValue CAMPFIRE_MAX_LIT_TIME_EXTENSION;
@@ -37,7 +36,6 @@ public class NeoForgeConfigHelper implements IConfigHelper {
     public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_RUN_OUT_INDICATOR_TIME;
     public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_RAIN_UNLIT_TIME;
     public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR;
-    public static final ModConfigSpec.BooleanValue SOUL_CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN;
     public static final ModConfigSpec.BooleanValue SOUL_CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME;
     public static final ModConfigSpec.BooleanValue SOUL_CAMPFIRE_ADDING_BURNABLES;
     public static final ModConfigSpec.IntValue SOUL_CAMPFIRE_MAX_LIT_TIME_EXTENSION;
@@ -65,10 +63,6 @@ public class NeoForgeConfigHelper implements IConfigHelper {
         CAMPFIRE_RAIN_PARTICLE_FACTOR = builder
                 .comment(CAMPFIRE_RAIN_PARTICLE_FACTOR_DESCRIPTION)
                 .defineInRange("campfireRainParticleFactor", CAMPFIRE_RAIN_PARTICLE_FACTOR_DEFAULT, CAMPFIRE_RAIN_PARTICLE_FACTOR_MIN, CAMPFIRE_RAIN_PARTICLE_FACTOR_MAX);
-
-        CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN = builder
-                .comment(CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN_DESCRIPTION)
-                .define("campfireDropsItemsWhenUnlitByTimeOrRain", CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN_DEFAULT);
 
         CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME = builder
                 .comment(CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME_DESCRIPTION)
@@ -109,10 +103,6 @@ public class NeoForgeConfigHelper implements IConfigHelper {
         SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR = builder
                 .comment(SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR_DESCRIPTION)
                 .defineInRange("soulCampfireRainParticleFactor", SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR_DEFAULT, SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR_MIN, SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR_MAX);
-
-        SOUL_CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN = builder
-                .comment(SOUL_CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN_DESCRIPTION)
-                .define("soulCampfireDropsItemsWhenUnlitByTimeOrRain", SOUL_CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN_DEFAULT);
 
         SOUL_CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME = builder
                 .comment(SOUL_CAMPFIRE_BREAKS_WHEN_UNLIT_BY_TIME_DESCRIPTION)
@@ -181,15 +171,6 @@ public class NeoForgeConfigHelper implements IConfigHelper {
             return isSoulCampfire ? SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR.get() : CAMPFIRE_RAIN_PARTICLE_FACTOR.get();
         } catch (IllegalStateException ex) {
             return isSoulCampfire ? SOUL_CAMPFIRE_RAIN_PARTICLE_FACTOR_DEFAULT : CAMPFIRE_RAIN_PARTICLE_FACTOR_DEFAULT;
-        }
-    }
-
-    @Override
-    public boolean isDroppingItemsWhenUnlitByTimeOrRain(boolean isSoulCampfire) {
-        try {
-            return isSoulCampfire ? SOUL_CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN.get() : CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN.get();
-        } catch (IllegalStateException ex) {
-            return isSoulCampfire ? SOUL_CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN_DEFAULT : CAMPFIRE_DROPS_ITEMS_WHEN_UNLIT_BY_TIME_OR_RAIN_DEFAULT;
         }
     }